#324 refactor : 종목 추가 버튼 위치 변경, 불필요한 PortfolioOverviewBodyM 삭제
