+++ src/features/portfolio/components/Portfolio/mobile/MainPanelM.tsx
@@ -1,9 +1,13 @@
+import Button from "@components/Buttons/Button";
+import { Icon } from "@components/Icon";
 import {
   PortfolioDetails,
   PortfolioHolding,
 } from "@features/portfolio/api/types";
+import { useBoolean } from "@fineants/demolition";
 import styled from "styled-components";
 import PortfolioHeaderM from "../../PortfolioHeaderM";
+import PortfolioHoldingAddDialog from "../../PortfolioHolding/PortfolioHoldingAddDialog";
 import PortfolioHoldingCardTable from "../../PortfolioHolding/mobile/PortfolioHoldingCardTable";
 import PortfolioOverviewM from "../../PortfolioOverview/mobile/PortfolioOverviewM";
 
@@ -20,6 +24,12 @@ export default function MainPanelM({
   tab,
   onChangeTab,
 }: Props) {
+  const {
+    state: isAddHoldingDialogOpen,
+    setTrue: onDialogOpen,
+    setFalse: onDialogClose,
+  } = useBoolean();
+
   return (
     <StyledMainPanel>
       <PortfolioHeaderM
@@ -32,6 +42,21 @@ export default function MainPanelM({
       <PortfolioOverviewContainer>
         <PortfolioOverviewM data={freshPortfolioDetailsData} />
       </PortfolioOverviewContainer>
+      {freshPortfolioHoldingsData.length !== 0 && (
+        <ButtonWrapper>
+          <Button variant="primary" size="h40" onClick={onDialogOpen}>
+            <Icon icon="add" size={16} color="white" />
+            <span>종목 추가</span>
+          </Button>
+        </ButtonWrapper>
+      )}
+
+      {isAddHoldingDialogOpen && (
+        <PortfolioHoldingAddDialog
+          isOpen={isAddHoldingDialogOpen}
+          onClose={onDialogClose}
+        />
+      )}
 
       <PortfolioHoldingCardTable data={freshPortfolioHoldingsData} />
     </StyledMainPanel>
@@ -44,4 +69,10 @@ const StyledMainPanel = styled.div`
 
 const PortfolioOverviewContainer = styled.div`
   width: 100%;
+`;
+
+const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: right;
+  padding: 16px 16px 24px;
 `;
